% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3graph.R
\name{d3graph}
\alias{d3graph}
\title{D3 Graph Interactive Widget}
\usage{
d3graph(g, width = NULL, height = NULL, elementId = NULL, show_labels = TRUE)
}
\arguments{
\item{g}{An igraph object}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{Element ID for the widget}

\item{show_labels}{Logical, whether to show node labels}
}
\description{
Create an interactive D3.js graph visualization
}
